refactor: 화면 로직에서 비즈니스 로직을 분리한다
